requirements.txt
streamlit>=1.28.0
swarms>=6.0.0
openai>=1.3.0
plotly>=5.17.0
pandas>=2.0.0
markdown2>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
loguru>=0.7.0
cryptography>=41.0.0
